+++ favoriter.html
@@ -8,8 +8,7 @@
     <link rel="stylesheet" href="css/style.css">
     <script src="js/hamburger.js"></script>
     <script src="js/karta.js"></script>
-    <script type="module" src="js/favoriter.js">
-    </script>
+    <script type="module" src="js/favoriter.js"></script>
     <script type="module" src="js/images.js"></script>
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" integrity="sha256-p4NxAoJBhIIN+hmNHrzRCf9tD/miZyoHS5obTRR9BMY=" crossorigin="">
     <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js" integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo=" crossorigin=""></script>
