+++ index.html
@@ -7,10 +7,8 @@
     <title>Namn - turista i Småland och på Öland</title>
     <link rel="stylesheet" href="css/style.css">
     <script type="module" src="js/script.js"></script>
-    <script src="js/test.js" defer></script>
-    <script src="js/formfunction.js"></script>
+    <script type="module" src="js/test.js" defer></script>
     <script src="js/hamburger.js"></script>
-    <script src="js/favoriter.js"></script>
     <script type="module" src="js/images.js"></script>
 
 </head>
@@ -80,8 +78,6 @@
             </section>
             <section>
                 <h2 id="popularDestination">Populära turistmål</h2>
-                <a href="#section2" class="scroll"><img src="photos/arrowdown.svg"></a>
-
             </section>
 
             <section id="section2">
