+++ js/test.js
@@ -1,3 +1,4 @@
+
 let questionElems = [
     document.querySelector("#question1"),
     document.querySelector("#question2"),
