+++ listsida.html
@@ -9,7 +9,6 @@
     <script src="js/hamburger.js"></script>
    <script type="module" src="/js/list.js"></script>
     <script src="js/karta.js"></script>
-    <script src="js/favoriter.js"></script>
   <script type="module" src="js/images.js"></script>
 
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" integrity="sha256-p4NxAoJBhIIN+hmNHrzRCf9tD/miZyoHS5obTRR9BMY=" crossorigin="">
